feat: Finalize workload generator

+++ src/workload_generator/uniform_generator.h
@@ -8,7 +8,7 @@ class UniformGenerator : public DistributionGenerator {
   std::uniform_int_distribution<uint64_t> dist;
 
 public:
-  UniformGenerator(uint64_t cardinality, uint32_t seed = 42)
+  UniformGenerator(uint64_t cardinality, uint64_t seed = 42)
       : gen(seed), dist(0, cardinality - 1) {}
 
   uint64_t Next() override { return dist(gen); }
